Fixed floating status bar on enemies sometimes not being set to current health on respawn
The InitializeFloatingStatusBar() was happening before the SetHealth/Mana/Stamina() functions so it was initializing to zero. The SetHealth/Mana/Stamina() functions were also happening before the FloatingStatusBar's Construct() in UMG was called so the AttributeListener task wasn't set up yet to get any changes.

The fix was to move the Respawn SetHealth/Mana/Stamina before InitializeFloatingStatusBar() so that InitializeFloatingStatusBar() initializes with the new values instead of relying on the AttributeListener to pickup the reset to max values on Respawn.