Changed ammo system to floats on Weapons + AmmoAttributeSet on Hero
Old system crashed if the ammo replicated after the Weapon's AttributeSet was removed. Also it wouldn't allow updating ammo (pickups) without the weapon being equipped.

Alternatively I could have created a new AttributeSet class per Weapon class that is always on the Hero, but this would prevent multiple weapons of a weapon class in the player's inventory. This would be fine for this project since I will only allow one weapon per class, but I want to be general as possible as an example. I'd also have to copy the clip values into floats on the weapons when they're dropped.

I could have one giant AttributeSet on the Hero that holds all the clip ammos as Attributes, but this would also only allow one weapon per weapon class in the player's inventory.

The best solution seems to be to use raw floats for clip ammo on the weapons themselves with reserve ammo for each ammo type on a new AmmoAttributeSet that's always on the Hero. However, this prevents the use of GameplayEffects to change the clip ammo so we essentially have to do prediction of clip ammo ourselves since the clip ammo is plain floats not on an AttributeSet/ASC. We had to override GSGameplayAbility's CheckCost() and ApplyCost() to work with these floats.

Some additional things like changing reticles made its way into this commit.