Fixed 2nd weapon in inventory being visible on first switch to 3p
The weapon was attached to the 3p mesh which was being set to visible and propagate the visibility to all its children.

Now weapon root is attached to Character root instead of 3p mesh. 1p weapon and 3p weapon are still individually attached to Character 1p mesh and 3p mesh.